- #12: Adding the war tiers.
