+++ src/data/hanzhong/hanzhong-cities.ts
@@ -1,0 +1,88 @@
+import type { HanzhongCityType } from '../../types/hanzhong-city-type';
+
+export const HANZHONG_CITIES: HanzhongCityType[] = [
+  {
+    id: 'hanzhong--city--yangping-pass',
+    name: 'Yangping Pass',
+    position: '(61,106)',
+    tacticalPoints: 2000000,
+    description: 'Increases Wei Officers damage to the garrison by 20%',
+    earnings: {
+      woodRate: 180,
+      grainsRate: 180,
+      ironRate: 180,
+    },
+  },
+  {
+    id: 'hanzhong--city--north-summit',
+    name: 'North Summit',
+    position: '(52,74)',
+    tacticalPoints: 2000000,
+    description: 'Increases Shu Officers damage to the garrison by 20%',
+    earnings: {
+      woodRate: 180,
+      grainsRate: 180,
+      ironRate: 180,
+    },
+  },
+  {
+    id: 'hanzhong--city--maming-pavilion',
+    name: 'Maming Pavilion',
+    position: '(90,63)',
+    tacticalPoints: 2000000,
+    description: 'Increases Wu Officers damage to the garrison by 20%',
+    earnings: {
+      woodRate: 180,
+      grainsRate: 180,
+      ironRate: 180,
+    },
+  },
+  {
+    id: 'hanzhong--city--mianyang-county',
+    name: 'Mianyang County',
+    position: '(99,96)',
+    tacticalPoints: 2000000,
+    description: 'Increases Other Officers damage to the garrison by 20%',
+    earnings: {
+      woodRate: 180,
+      grainsRate: 180,
+      ironRate: 180,
+    },
+  },
+  {
+    id: 'hanzhong--city--mount-dingjun',
+    name: 'Mount Dingjun',
+    position: '(63,81)',
+    tacticalPoints: 5000000,
+    description: "Reduces the garrison's damage taken by 20%",
+    earnings: {
+      woodRate: 360,
+      grainsRate: 360,
+      ironRate: 360,
+    },
+  },
+  {
+    id: 'hanzhong--city--hanshui-trail',
+    name: 'Hanshui Trail',
+    position: '(87,91)',
+    tacticalPoints: 5000000,
+    description: "Reduces the garrison's damage taken by 20%",
+    earnings: {
+      woodRate: 360,
+      grainsRate: 360,
+      ironRate: 360,
+    },
+  },
+  {
+    id: 'hanzhong--city--hanzhong-city',
+    name: 'Hanzhong City',
+    position: '(75,85)',
+    tacticalPoints: 140000000,
+    description: "Reduces the garrison's damage taken by 20%",
+    earnings: {
+      woodRate: 720,
+      grainsRate: 720,
+      ironRate: 720,
+    },
+  },
+];
