+++ src/data/hanzhong/index.ts
@@ -1,0 +1,10 @@
+import { HANZHONG_CITIES } from './hanzhong-cities';
+import { HANZHONG_TERRITORIES } from './hanzhong-territories';
+import { HANZHONG_WAR_TIERS } from './hanzhong-war-tiers';
+import type { HanzhongType } from '../../types/hanzhong-type';
+
+export const HANZHONG_DATA: HanzhongType = {
+  cities: HANZHONG_CITIES,
+  territories: HANZHONG_TERRITORIES,
+  warTiers: HANZHONG_WAR_TIERS,
+};
